Update the message handling event
The server was updated to send two separate events for a correct answer and a regular message. In case of the former, the server now sends a 'correct' event instead of a 'message' event accompanied with a type. As for regular messages, no modifications were brought. Therefore, the client has been updated to handle the 'correct' event